+++ FixedspawnDissonance/code/Artifacts/Glass.cs
@@ -14,17 +14,17 @@ namespace FixedspawnDissonance
         {
             IL.RoR2.CharacterModel.UpdateOverlays += CharacterModel_UpdateOverlayStates;
             IL.RoR2.CharacterModel.UpdateOverlayStates += CharacterModel_UpdateOverlayStates;
-
-            
+ 
         }
 
         private static void CharacterModel_UpdateOverlayStates(ILContext il)
         {
             ILCursor c = new ILCursor(il);
-            if (c.TryGotoNext(MoveType.Before,
-            x => x.MatchLdsfld("RoR2.RoR2Content/Items", "LunarDagger")))
+            bool a = c.TryGotoNext(MoveType.Before,
+            x => x.MatchLdsfld("RoR2.RoR2Content/Items", "LunarDagger"));
+            if (a && c.TryGotoNext(MoveType.Before,
+            x => x.MatchBr(out _)))
             {
-                c.Index += 4;
                 c.Emit(OpCodes.Ldarg_0);
                 c.EmitDelegate<System.Func<bool, CharacterModel, bool>>((yes, model) =>
                 {